requirements-dev.txt
Django
django-parler
rdflib
pytest
pytest-cov
pytest-django
pytest-factoryboy
flake8
isort
pep8-naming
